SW: Basic AES capture script works.
 - Ran 2to3 on most things.
 - Fix some string handling (will need to test carefully).
 - Started deleting GUI stuff. Lots more to come.

+++ software/chipwhisperer/common/ui/ProgressBar.py
@@ -89,7 +89,7 @@ class ProgressBarText(object):
         spent = (datetime.now() - self.startTime).total_seconds()
         eta = spent * (self.maximum - self.currentProgress) / self.currentProgress
         try:
-            eta = timedelta(seconds=long(eta))
+            eta = timedelta(seconds=int(eta))
         except OverflowError:
             return "n/a"
         if eta.days > 9:
